% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalproject.R
\name{majoritystudents}
\alias{majoritystudents}
\title{majority students}
\usage{
majoritystudents()
}
\value{
subset of size data frame where only remaining data is the specified type of student population
}
\description{
subset by majority undergraduate/graduate students
}
