% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalproject.R
\name{region}
\alias{region}
\title{region}
\usage{
region()
}
\value{
subset of original data frame where only remaining are those within the user specified region
}
\description{
subset by region of school
}
