% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalproject.R
\name{scores}
\alias{scores}
\title{test score}
\usage{
scores()
}
\value{
subset of price data frame only including schools where student score above midline on ACT or SAT
}
\description{
subset by test scores
}
